// src/app/todos/models/task.models.ts
export interface GetTasksResponse {
  items: Task[]
  totalCount: number
  error: string
}

export interface Task extends UpdateTaskModel {
  id: string
  todoListId: string
  order: number
  addedDate: string
}

export interface UpdateTaskModel {
  title: string
  description: string
  completed: boolean
  status: number
  priority: number
  startDate: string
  deadline: string
}

// {
//   todoId: [] (Task[]),
//   todoId2: [] (Task[]),
// }

export interface DomainTask {
  [key: string]: Task[]
}